try operations abstract base class
