+++ connexion/operations/__init__.py
@@ -1,0 +1,4 @@
+from .abstract import AbstractOperation  # noqa
+from .openapi import OpenAPIOperation  # noqa
+from .swagger2 import Swagger2Operation  # noqa
+from .secure import SecureOperation  # noqa
