+++ tests/test_mock.py
@@ -1,5 +1,5 @@
 from connexion.mock import MockResolver, partial
-from connexion.operation import Operation, Swagger2Operation
+from connexion.operations import OpenAPIOperation, Swagger2Operation
 
 
 def test_partial():
@@ -27,7 +27,7 @@ def test_mock_resolver():
         }
     }
 
-    operation = Operation(api=None,
+    operation = OpenAPIOperation(api=None,
                           method='GET',
                           path='endpoint',
                           path_parameters=[],
@@ -87,7 +87,7 @@ def test_mock_resolver_ref_schema_example():
         }
     }
 
-    operation = Operation(api=None,
+    operation = OpenAPIOperation(api=None,
                           method='GET',
                           path='endpoint',
                           path_parameters=[],
@@ -136,7 +136,7 @@ def test_mock_resolver_inline_schema_example():
         }
     }
 
-    operation = Operation(api=None,
+    operation = OpenAPIOperation(api=None,
                           method='GET',
                           path='endpoint',
                           path_parameters=[],
@@ -158,7 +158,7 @@ def test_mock_resolver_no_examples():
         '418': {}
     }
 
-    operation = Operation(api=None,
+    operation = OpenAPIOperation(api=None,
                           method='GET',
                           path='endpoint',
                           path_parameters=[],
@@ -182,7 +182,7 @@ def test_mock_resolver_notimplemented():
     }
 
     # do not mock the existent functions
-    operation = Operation(api=None,
+    operation = OpenAPIOperation(api=None,
                           method='GET',
                           path='endpoint',
                           path_parameters=[],
@@ -194,7 +194,7 @@ def test_mock_resolver_notimplemented():
     assert operation.operation_id == 'fakeapi.hello.get'
 
     # mock only the nonexistent ones
-    operation = Operation(api=None,
+    operation = OpenAPIOperation(api=None,
                           method='GET',
                           path='endpoint',
                           path_parameters=[],
@@ -206,4 +206,4 @@ def test_mock_resolver_notimplemented():
                           resolver=resolver)
 
     # check if it is using the mock function
-    assert operation._undecorated_function() == ('No example response was defined.', 418)
+    assert operation._resolution.function() == ('No example response was defined.', 418)
