+++ tests/test_resolver.py
@@ -1,7 +1,7 @@
 import connexion.apps
 import pytest
 from connexion.exceptions import ResolverError
-from connexion.operation import Operation, Swagger2Operation
+from connexion.operations import OpenAPIOperation
 from connexion.resolver import Resolver, RestyResolver
 
 COMPONENTS = {'parameters': {'myparam': {'in': 'path', 'schema': {'type': 'integer'}}}}
@@ -36,7 +36,7 @@ def test_bad_operation_id():
 
 
 def test_standard_resolve_x_router_controller():
-    operation = Operation(api=None,
+    operation = OpenAPIOperation(api=None,
                           method='GET',
                           path='endpoint',
                           path_parameters=[],
@@ -51,7 +51,7 @@ def test_standard_resolve_x_router_controller():
 
 
 def test_resty_resolve_operation_id():
-    operation = Operation(api=None,
+    operation = OpenAPIOperation(api=None,
                           method='GET',
                           path='endpoint',
                           path_parameters=[],
@@ -65,7 +65,7 @@ def test_resty_resolve_operation_id():
 
 
 def test_resty_resolve_x_router_controller_with_operation_id():
-    operation = Operation(api=None,
+    operation = OpenAPIOperation(api=None,
                           method='GET',
                           path='endpoint',
                           path_parameters=[],
@@ -80,7 +80,7 @@ def test_resty_resolve_x_router_controller_with_operation_id():
 
 
 def test_resty_resolve_x_router_controller_without_operation_id():
-    operation = Operation(api=None,
+    operation = OpenAPIOperation(api=None,
                           method='GET',
                           path='/hello/{id}',
                           path_parameters=[],
@@ -92,7 +92,7 @@ def test_resty_resolve_x_router_controller_without_operation_id():
 
 
 def test_resty_resolve_with_default_module_name():
-    operation = Operation(api=None,
+    operation = OpenAPIOperation(api=None,
                           method='GET',
                           path='/hello/{id}',
                           path_parameters=[],
@@ -104,7 +104,7 @@ def test_resty_resolve_with_default_module_name():
 
 
 def test_resty_resolve_with_default_module_name_lowercase_verb():
-    operation = Operation(api=None,
+    operation = OpenAPIOperation(api=None,
                           method='get',
                           path='/hello/{id}',
                           path_parameters=[],
@@ -116,7 +116,7 @@ def test_resty_resolve_with_default_module_name_lowercase_verb():
 
 
 def test_resty_resolve_with_default_module_name_will_translate_dashes_in_resource_name():
-    operation = Operation(api=None,
+    operation = OpenAPIOperation(api=None,
                           method='GET',
                           path='/foo-bar',
                           path_parameters=[],
@@ -128,7 +128,7 @@ def test_resty_resolve_with_default_module_name_will_translate_dashes_in_resourc
 
 
 def test_resty_resolve_with_default_module_name_can_resolve_api_root():
-    operation = Operation(api=None,
+    operation = OpenAPIOperation(api=None,
                           method='GET',
                           path='/',
                           path_parameters=[],
@@ -140,7 +140,7 @@ def test_resty_resolve_with_default_module_name_can_resolve_api_root():
 
 
 def test_resty_resolve_with_default_module_name_will_resolve_resource_root_get_as_search():
-    operation = Operation(api=None,
+    operation = OpenAPIOperation(api=None,
                           method='GET',
                           path='/hello',
                           path_parameters=[],
@@ -152,7 +152,7 @@ def test_resty_resolve_with_default_module_name_will_resolve_resource_root_get_a
 
 
 def test_resty_resolve_with_default_module_name_and_x_router_controller_will_resolve_resource_root_get_as_search():
-    operation = Operation(api=None,
+    operation = OpenAPIOperation(api=None,
                           method='GET',
                           path='/hello',
                           path_parameters=[],
@@ -166,7 +166,7 @@ def test_resty_resolve_with_default_module_name_and_x_router_controller_will_res
 
 
 def test_resty_resolve_with_default_module_name_will_resolve_resource_root_as_configured():
-    operation = Operation(api=None,
+    operation = OpenAPIOperation(api=None,
                           method='GET',
                           path='/hello',
                           path_parameters=[],
@@ -178,7 +178,7 @@ def test_resty_resolve_with_default_module_name_will_resolve_resource_root_as_co
 
 
 def test_resty_resolve_with_default_module_name_will_resolve_resource_root_post_as_post():
-    operation = Operation(api=None,
+    operation = OpenAPIOperation(api=None,
                           method='POST',
                           path='/hello',
                           path_parameters=[],
